Remove nestedjsonparser
