+++ sycamore/scans/__init__.py
@@ -1,4 +1,4 @@
-from sycamore.scans.file_scan import BinaryScan, FileScan, JsonScan, NestedJsonScan
+from sycamore.scans.file_scan import BinaryScan, FileScan, JsonScan
 from sycamore.scans.materialized_scan import ArrowScan, DocScan, MaterializedScan, PandasScan
 
 __all__ = [
@@ -7,7 +7,6 @@ __all__ = [
     "DocScan",
     "FileScan",
     "JsonScan",
-    "NestedJsonScan",
     "MaterializedScan",
     "PandasScan",
 ]
